refactor: components to home
