+++ src/presentation/pages/Home/components/NavBar/style.ts
@@ -1,0 +1,16 @@
+import { styled } from 'styled-components'
+
+export const Nav = styled.nav`
+  width: 100%;
+  background: var(--white);
+  display: flex;
+  gap: 48%;
+  align-items: center;
+  height: auto;
+  top: 0;
+  padding: 8px 16px;
+
+  @media (max-width: 768px) {
+    gap: 16%;
+  }
+`
